specification: shorten slightly the text
With the new forced page breaks (previous commit) two pages had just a
single line in them. So this shrinks a little the text to fix the
problem.  I don't want to force page breaks only in some of the
subsections, for consistency.

Note: I tried the argument "3" to page breaks, to have them strongly
"suggeested" but not "mandated" (argument 4 or default if no argument),
but the result was like without page breaks (i.e., bad).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>