specification: explain why we don't support interrupts in SDB
This adds a section that explains why we don't describe interrupts.
Most of the section comes from a message by Wesley Terpstra,

  http://lists.ohwr.org/sympa/arc/fpga-config-space/2013-03/msg00000.html

(you may need to re-enter the same url in the browser after clicking on
the "I'm not a spammer" button).

I'm not personally able to make the discussion shorter, because I think
it's very informative.  The message has been slightly reworded to make
it less informal and more suitable to be an "official" rationale, and
I added some glue lines at the beginning and the end.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>