vme-wb: add the wb to vme bridge kernel module.
It allows to use from a vme master eb commands (e.g. eb-ls) over the vme interface.
It follows the philosophy of pci-wb. So far only reading eb commands are supported.
The development is in a test phase, for that reason there are hardcode values and paths.

Main files:
vme_wb.*: register the kernel module and creates two VME windows, the first for the
          setting of the CSR/CR parameters in the vme slave, and second for data
          communication.
Module.symvers: it is needed for the correct compilation of the module sice it uses functions
                from wishbone.ko and vmebus.ko.
scripts for inserting and removing the module.