pcie/vme bug: race condition in checking empty flag and pop
The drivers were popping the slave request FIFO, whether it was empty or not.
This seemed safe since empty FIFOs do nothing when popped. However, if the
FIFO receives an element between the empty check and the pop, we might pop
an element without ever seeing it. Under load this led to lost interrupts.