wishbone: remove mutex use from interrupt handler
Previously, the interrupt handler popped the data from the hardware.
To do this, it needed to lock the target buffers, guarded by a mutex.
Unfortunately, that can cause the interrupt handler to sleep! bad!

This patch moves the waitq and fasync structure to the wishbone structure
which is always present.  Then, the interrupt handler need only wake up any
threads blocked waiting for data.  The read/poll functions invoke the
processing of data from the hardware, instead.