sdbfs/lib: add flag ZEROBASED and remove verbose argument to register()
Adding ZEROBASED allows the lm32 to enumerate fpga cores without
prviding a specific read() method.

Removing "verbose" as an argument to sdbfs_dev_create (and allowing
the user to set it in fs->flags) is a minor cleanup, though incompatible.
Userspace is fixed to deal with this incompatibility.

This is a slight incompatibility, but it's worh doing.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>

s